requirements.txt
crewai[tools]>=0.108.0
litellm
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
rich>=13.9.0
pyyaml>=6.0.0
# Interfaces
pyTelegramBotAPI>=4.25.0
streamlit>=1.42.0
groq>=0.12.0
gTTS>=2.5.0
sounddevice>=0.5.0
